+++ packages/quick_actions/quick_actions_platform_interface/lib/types/shortcut_item.dart
@@ -11,6 +11,7 @@ class ShortcutItem {
   const ShortcutItem({
     required this.type,
     required this.localizedTitle,
+    this.localizedSubtitle,
     this.icon,
   });
 
@@ -19,6 +20,11 @@ class ShortcutItem {
 
   /// Localized title of the item.
   final String localizedTitle;
+
+  /// Localized subtitle of the item.
+  ///
+  /// This is ignored on Android, since it's not supported.
+  final String? localizedSubtitle;
 
   /// Name of native resource (xcassets etc; NOT a Flutter asset) to be
   /// displayed as the icon for this item.