+++ packages/quick_actions/quick_actions_ios/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Add localizedSubtitle field
+
 ## 1.1.1
 
 * Updates to a newer version of Pigeon.